fix: added jsdocs
